New rooms are now saved in database instead of in csv file.

+++ src/main/java/org/example/domain/room/RoomFileRepository.java
@@ -30,7 +30,7 @@ public class RoomFileRepository implements RoomRepository {
         return newRoom;
     }
 
-    Room addExistingRoom(int id, int number, List<BedType> bedTypes) {
+    Room addExistingRoom(long id, int number, List<BedType> bedTypes) {
         Room newRoom = new Room(id, number, bedTypes);
         rooms.add(newRoom);
         return newRoom;
@@ -97,8 +97,8 @@ public class RoomFileRepository implements RoomRepository {
         }
     }
 
-    private int findNewId() {
-        int max = 0;
+    private long findNewId() {
+        long max = 0;
         for (Room room : rooms) {
             if (room.getId() > max) {
                 max = room.getId();
@@ -108,7 +108,7 @@ public class RoomFileRepository implements RoomRepository {
     }
 
     @Override
-    public void remove(int id) {
+    public void remove(long id) {
         int roomToBeRemovedIndex = -1;
         for (int i = 0; i < this.rooms.size(); i++) {
             if (this.rooms.get(i).getId() == id) {
@@ -122,13 +122,13 @@ public class RoomFileRepository implements RoomRepository {
     }
 
     @Override
-    public void edit(int id, int number, List<BedType> bedTypes) {
+    public void edit(long id, int number, List<BedType> bedTypes) {
         this.remove(id);
         this.addExistingRoom(id, number, bedTypes);
     }
 
     @Override
-    public Room getById(int id) {
+    public Room getById(long id) {
         for (Room room : rooms) {
             if (room.getId() == id) {
                 return room;
